#XNX-9 | Implemented persistent storage for game levels #XNX-4 | Added three demo levels #XNX-8 | Implemented initialisation of enemy and protagonist positions from level file

+++ Main/src/net/ivang/xonix/main/Enemy.java
@@ -13,7 +13,7 @@ public class Enemy {
     private Move moveDirection;
     private GameMap gameMap;
 
-    Enemy(int x, int y, GameMap gameMap) {
+    Enemy(float x, float y, GameMap gameMap) {
         pos = new Vector2(x, y);
         moveDirection = Move.UP_LEFT;
         this.gameMap = gameMap;
@@ -93,4 +93,5 @@ public class Enemy {
     public void setMoveDirection(Move moveDirection) {
         this.moveDirection = moveDirection;
     }
+
 }
